+++ src/main/java/net/geforcemods/securitycraft/blocks/BlockLaserBlock.java
@@ -4,13 +4,14 @@ import java.util.Random;
 
 import net.geforcemods.securitycraft.ConfigHandler;
 import net.geforcemods.securitycraft.SCContent;
-import net.geforcemods.securitycraft.api.CustomizableSCTE;
+import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.api.IOwnable;
-import net.geforcemods.securitycraft.api.Owner;
-import net.geforcemods.securitycraft.api.TileEntityLinkable;
 import net.geforcemods.securitycraft.misc.EnumModuleType;
+import net.geforcemods.securitycraft.network.client.OpenLaserScreen;
 import net.geforcemods.securitycraft.tileentity.TileEntityLaserBlock;
 import net.geforcemods.securitycraft.util.BlockUtils;
+import net.geforcemods.securitycraft.util.PlayerUtils;
+import net.geforcemods.securitycraft.util.Utils;
 import net.minecraft.block.Block;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
@@ -19,12 +20,15 @@ import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumBlockRenderType;
 import net.minecraft.util.EnumFacing;
+import net.minecraft.util.EnumHand;
 import net.minecraft.util.EnumParticleTypes;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
@@ -49,56 +53,86 @@ public class BlockLaserBlock extends BlockDisguisable {
 	}
 
 	@Override
+	public boolean onBlockActivated(World world, BlockPos pos, IBlockState state, EntityPlayer player, EnumHand hand, EnumFacing facing, float hitX, float hitY, float hitZ) {
+		TileEntityLaserBlock be = (TileEntityLaserBlock) world.getTileEntity(pos);
+
+		if (!world.isRemote && be.isOwnedBy(player)) {
+			if (!be.isEnabled())
+				player.sendStatusMessage(Utils.localize("gui.securitycraft:scManual.disabled"), true);
+			else
+				SecurityCraft.network.sendTo(new OpenLaserScreen(pos, be.getSideConfig()), (EntityPlayerMP) player);
+		}
+
+		return true;
+	}
+
+	@Override
 	public void onBlockPlacedBy(World world, BlockPos pos, IBlockState state, EntityLivingBase entity, ItemStack stack) {
 		super.onBlockPlacedBy(world, pos, state, entity, stack);
 
-		if (!world.isRemote && entity instanceof EntityPlayer)
-			setLaser(new Owner(entity.getName(), entity.getUniqueID().toString()), world, pos);
+		if (!world.isRemote)
+			setLaser(world, pos, entity instanceof EntityPlayer ? (EntityPlayer) entity : null);
 	}
 
-	public void setLaser(Owner owner, World world, BlockPos pos) {
+	public void setLaser(World level, BlockPos pos, EntityPlayer player) {
+		for (EnumFacing facing : EnumFacing.values()) {
+			setLaser(level, pos, facing, player);
+		}
+	}
+
+	public void setLaser(World world, BlockPos pos, EnumFacing facing, EntityPlayer player) {
 		TileEntityLaserBlock thisTe = (TileEntityLaserBlock) world.getTileEntity(pos);
 
-		for (EnumFacing facing : EnumFacing.VALUES) {
-			int boundType = facing == EnumFacing.UP || facing == EnumFacing.DOWN ? 1 : (facing == EnumFacing.NORTH || facing == EnumFacing.SOUTH ? 2 : 3);
+		if (!thisTe.isSideEnabled(facing))
+			return;
 
-			inner:
-			for (int i = 1; i <= ConfigHandler.laserBlockRange; i++) {
-				BlockPos offsetPos = pos.offset(facing, i);
-				IBlockState offsetState = world.getBlockState(offsetPos);
-				Block offsetBlock = offsetState.getBlock();
+		int boundType = facing == EnumFacing.UP || facing == EnumFacing.DOWN ? 1 : (facing == EnumFacing.NORTH || facing == EnumFacing.SOUTH ? 2 : 3);
 
-				if (!offsetBlock.isAir(offsetState, world, offsetPos) && !offsetBlock.isReplaceable(world, offsetPos) && offsetBlock != SCContent.laserBlock)
-					break inner;
-				else if (offsetBlock == SCContent.laserBlock) {
-					TileEntityLaserBlock thatTe = (TileEntityLaserBlock) world.getTileEntity(offsetPos);
+		for (int i = 1; i <= ConfigHandler.laserBlockRange; i++) {
+			BlockPos offsetPos = pos.offset(facing, i);
+			IBlockState offsetState = world.getBlockState(offsetPos);
+			Block offsetBlock = offsetState.getBlock();
 
-					if (owner.owns(thatTe)) {
-						TileEntityLinkable.link(thisTe, thatTe);
+			if (!offsetBlock.isAir(offsetState, world, offsetPos) && !offsetBlock.isReplaceable(world, offsetPos) && offsetBlock != SCContent.laserBlock)
+				return;
+			else if (offsetBlock == SCContent.laserBlock) {
+				TileEntityLaserBlock thatTe = (TileEntityLaserBlock) world.getTileEntity(offsetPos);
 
-						for (EnumModuleType type : thatTe.getInsertedModules()) {
-							thisTe.insertModule(thatTe.getModule(type), false);
-						}
-
-						if (thisTe.isEnabled() && thatTe.isEnabled()) {
-							for (int j = 1; j < i; j++) {
-								offsetPos = pos.offset(facing, j);
-								offsetState = world.getBlockState(offsetPos);
-
-								if (offsetState.getBlock().isAir(offsetState, world, offsetPos) || offsetState.getBlock().isReplaceable(world, offsetPos)) {
-									world.setBlockState(offsetPos, SCContent.laserField.getDefaultState().withProperty(BlockLaserField.BOUNDTYPE, boundType));
-
-									TileEntity te = world.getTileEntity(offsetPos);
-
-									if (te instanceof IOwnable)
-										((IOwnable) te).setOwner(thisTe.getOwner().getUUID(), thisTe.getOwner().getName());
-								}
-							}
-						}
+				if (thisTe.getOwner().owns(thatTe) && thisTe.isEnabled() && thatTe.isEnabled()) {
+					if (!thatTe.isSideEnabled(facing.getOpposite())) {
+						thisTe.setSideEnabled(facing, false, null);
+						return;
 					}
 
-					break inner;
+					EnumModuleType failedType = thisTe.synchronizeWith(thatTe);
+
+					if (failedType != null) {
+						if (player != null) {
+							PlayerUtils.sendMessageToPlayer(player, Utils.localize(this), Utils.localize("messages.securitycraft:laser.sync_failed", Utils.getFormattedCoordinates(thatTe.getPos()), Utils.localize(failedType.getTranslationKey())), TextFormatting.RED);
+							thisTe.setSideEnabled(facing, false, null);
+							thatTe.setSideEnabled(facing.getOpposite(), false, null);
+							player.closeScreen();
+						}
+
+						return;
+					}
+
+					for (int j = 1; j < i; j++) {
+						offsetPos = pos.offset(facing, j);
+						offsetState = world.getBlockState(offsetPos);
+
+						if (offsetState.getBlock().isAir(offsetState, world, offsetPos) || offsetState.getBlock().isReplaceable(world, offsetPos)) {
+							world.setBlockState(offsetPos, SCContent.laserField.getDefaultState().withProperty(BlockLaserField.BOUNDTYPE, boundType));
+
+							TileEntity te = world.getTileEntity(offsetPos);
+
+							if (te instanceof IOwnable)
+								((IOwnable) te).setOwner(thisTe.getOwner().getUUID(), thisTe.getOwner().getName());
+						}
+					}
 				}
+
+				return;
 			}
 		}
 	}
@@ -110,12 +144,12 @@ public class BlockLaserBlock extends BlockDisguisable {
 	}
 
 	public static void destroyAdjacentLasers(World world, BlockPos pos) {
-		BlockUtils.destroyInSequence(SCContent.laserField, world, pos, EnumFacing.VALUES);
+		BlockUtils.removeInSequence(SCContent.laserField, world, pos, EnumFacing.VALUES);
 	}
 
 	@Override
 	public void neighborChanged(IBlockState state, World world, BlockPos pos, Block block, BlockPos fromPos) {
-		setLaser(((CustomizableSCTE) world.getTileEntity(pos)).getOwner(), world, pos);
+		setLaser(world, pos, null);
 	}
 
 	@Override